backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "egame-sim"
version = "0.1.0"
description = "AI-driven evolutionary sandbox backend"
authors = [{ name = "Codex" }]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.21",
    "pydantic-settings>=2.2.1",
    "numpy>=2.0.0",
    "scipy>=1.11.0",
    "networkx>=3.2.1",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "anyio>=4.4.0",
    "imageio>=2.31.0",
    "lz4>=4.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

